+++ src/imio/project/pst/migrations/migrate_to_1_3_1.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from collective.documentgenerator.utils import update_oo_config
+from collective.messagesviewlet.utils import add_message
 from eea.facetednavigation.criteria.interfaces import ICriteria
 from eea.facetednavigation.subtypes.interfaces import IFacetedNavigable
 from eea.facetednavigation.widgets.storage import Criterion
@@ -23,6 +24,45 @@ class Migrate_To_1_3_1(Migrator):
     def __init__(self, context):
         Migrator.__init__(self, context)
 
+    def run(self):
+        # check if oo port must be changed
+        update_oo_config()
+
+        self.runProfileSteps('imio.project.pst', steps=['actions', 'catalog'], run_dependencies=False)
+
+        # migrate projectspace in pstprojectspace
+        self.migrate_projectspace_in_pstprojectspace()
+
+        # Assigning new custom ecomptes permissions to role
+        self.manage_permission()
+
+        # update daterange criteria
+        self.update_dashboards()
+
+        # remove configlets
+        self.remove_configlets()
+
+        # Allowed webservices on subactions
+        self.configure_webservices()
+
+        # reindex all actions
+        self.reindex_all_actions()
+
+        # upgrade all except 'imio.project.pst:default'. Needed with bin/upgrade-portals
+        self.upgradeAll(omit=['imio.project.pst:default'])
+
+        # for prod in []:
+        #     mark_last_version(self.portal, product=prod)
+
+        # Reorder css and js
+        self.runProfileSteps('imio.project.pst', steps=['cssregistry', 'jsregistry'], run_dependencies=False)
+
+        # Add a new-version warning message in message config
+        self.add_new_version_message()
+
+        # Display duration
+        self.finish()
+
     def add_plan_to_lists_fields(self, *lists_fields):
         for list_fields in lists_fields:
             if 'plan' not in list_fields:
@@ -30,6 +70,16 @@ class Migrate_To_1_3_1(Migrator):
                     list_fields.insert(list_fields.index('categories') + 1, 'plan')
                 else:
                     list_fields.append('plan')
+
+    def migrate_pst_fields(self, field_list):
+        updated_list = []
+        for field_name in field_list:
+            updated_list.append({
+                'field_name': field_name,
+                'read_tal_condition': '',
+                'write_tal_condition': '',
+            })
+        return updated_list
 
     def update_dashboards(self):
         # update daterange criteria
@@ -51,12 +101,7 @@ class Migrate_To_1_3_1(Migrator):
                 criterion.criteria[position] = Criterion(**values)
                 criterion.criteria._p_changed = 1
 
-    def run(self):
-        # check if oo port must be changed
-        update_oo_config()
-
-        self.runProfileSteps('imio.project.pst', steps=['actions', 'catalog'], run_dependencies=False)
-
+    def migrate_projectspace_in_pstprojectspace(self):
         plan_values = [
             {'label': u"Agenda 21 local",
              'key': "agenda-21-local"},
@@ -109,11 +154,11 @@ class Migrate_To_1_3_1(Migrator):
                         new_class_name='imio.project.pst.content.pstprojectspace.PSTProjectSpace')
                     # projectspace is now pstprojectspace
                     projectspace_obj.portal_type = 'pstprojectspace'
-                    projectspace_obj.project_fields = prj_fld_record
-                    projectspace_obj.strategicobjective_fields = so_record
-                    projectspace_obj.operationalobjective_fields = oo_record
-                    projectspace_obj.pstaction_fields = a_record
-                    projectspace_obj.pstsubaction_fields = sa_record
+                    projectspace_obj.project_fields = self.migrate_pst_fields(prj_fld_record)
+                    projectspace_obj.strategicobjective_fields = self.migrate_pst_fields(so_record)
+                    projectspace_obj.operationalobjective_fields = self.migrate_pst_fields(oo_record)
+                    projectspace_obj.pstaction_fields = self.migrate_pst_fields(a_record)
+                    projectspace_obj.pstsubaction_fields = self.migrate_pst_fields(sa_record)
                     if not getattr(projectspace_obj, 'plan_values'):
                         setattr(projectspace_obj, 'plan_values', plan_values)
                     if not getattr(projectspace_obj, 'strategicobjective_budget_states'):
@@ -131,21 +176,19 @@ class Migrate_To_1_3_1(Migrator):
                 del registry.records['imio.project.settings.pstaction_fields']
                 del registry.records['imio.project.settings.pstsubaction_fields']
 
-        # Assigning custom permission to role
+    def manage_permission(self):
         self.portal.manage_permission('imio.project.pst: ecomptes import',
                                       ('Manager', 'Site Administrator', 'Contributor'), acquire=0)
         self.portal.manage_permission('imio.project.pst: ecomptes export',
                                       ('Manager', 'Site Administrator', 'Contributor'), acquire=0)
 
-        # update daterange criteria
-        self.update_dashboards()
-
-        # remove configlets
+    def remove_configlets(self):
         config_tool = api.portal.get_tool('portal_controlpanel')
         config_tool.unregisterConfiglet('imio.project.core.settings')
         config_tool.unregisterConfiglet('imio.project.pst.settings')
 
-        # Allowed  web service actions on subactions
+    def configure_webservices(self):
+        registry = getUtility(IRegistry)
         generated_actions = registry.get('imio.pm.wsclient.browser.settings.IWS4PMClientSettings.generated_actions')
         for action in generated_actions:
             if action['condition'] == u"python: context.getPortalTypeName() in ('pstaction', 'task')":
@@ -156,21 +199,25 @@ class Migrate_To_1_3_1(Migrator):
         api.portal.set_registry_record('imio.pm.wsclient.browser.settings.IWS4PMClientSettings.generated_actions',
                                        generated_actions)
 
-        # reindex all actions
+    def reindex_all_actions(self):
         actions_brains = self.catalog(object_provides=IPSTAction.__identifier__)
         for action_brain in actions_brains:
             action_brain.getObject().reindexObject()
 
-        self.upgradeAll(omit=['imio.project.pst:default'])
-
-        for prod in []:
-            mark_last_version(self.portal, product=prod)
-
-        # Reorder css and js
-        self.runProfileSteps('imio.project.pst', steps=['cssregistry', 'jsregistry'], run_dependencies=False)
-
-        # Display duration
-        self.finish()
+    def add_new_version_message(self):
+        if 'new-version' in self.portal['messages-config']:
+            api.content.delete(self.portal['messages-config']['new-version'])
+        add_message(
+            'new-version',
+            'Version 1.3.1',
+            u'<p>Vous êtes passés à la version d\'iA.PST 1.3.1 !</p>'
+            u'<p>La <a href="https://docs.imio.be/imio-doc/ia.pst/" target="_blank">'
+            u'documentation</a> a été mise à jour et comporte une nouvelle section sur les nouveautés</a>.</p>',
+            msg_type='warning',
+            can_hide=True,
+            req_roles=['Authenticated'],
+            activate=True
+        )
 
 
 def migrate(context):